global: s3 refactoring
* Changes the file structure.

* Adds 100% coverage.
