+++ invenio_files_rest/storage.py
@@ -28,14 +28,9 @@ from __future__ import absolute_import, print_function
 
 import hashlib
 import time
-import tempfile
 from os.path import join
 
 from fs.opener import opener
-from boto.s3.connection import S3Connection
-from flask import current_app, redirect
-
-from boto.s3.key import Key
 
 from .errors import StorageError
 from .helpers import send_stream
@@ -132,83 +127,3 @@ class PyFilesystemStorage(Storage):
                 content_md5=self.file.checksum)
         except Exception as e:
             raise StorageError('Could not send file: {}'.format(e))
-
-
-class AmazonS3(Storage):
-    def __init__(self, object, fileinstance):
-        super(AmazonS3, self).__init__(object, fileinstance)
-        if "AWS_KEY" in current_app.config:
-            raise StorageError("Missing AWS Key")
-        if "AWS_SECRET" in current_app.config:
-            raise StorageError("Missing AWS Secret")
-        self.aws_connection = S3Connection(current_app.config.get("AWS_KEY", "AWS_KEY"),
-                                           current_app.config.get("AWS_SECRET", "AWS_SECRET"))
-
-    def get_size(self):
-        bucket_name, key_name = self.obj.file.uri.replace("s3:", "").split(":")
-        bucket = self.aws_connection.get_bucket(bucket_name)
-        return bucket.lookup(key_name).size
-
-    def open(self):
-
-        temp = tempfile.NamedTemporaryFile()
-        try:
-            loc = self.obj.bucket.location.uri
-            bucket = self.aws_connection.get_bucket(loc)
-            element = Key(bucket)
-            key_name = self.obj.file.uri.replace("s3:", "").split(":")[1]
-            element.key = key_name
-            element.get_contents_to_file(temp)
-            temp.seek(0)
-            return temp
-        except:
-            temp.close()
-            return None
-
-    def send_file(self, filename):
-        """Send the file to the client.
-
-        This returns a flask response that will eventually result in
-        the user being offered to download the file (or view it in the
-        browser).  Depending on the storage backend it may actually
-        send a redirect to an external URL where the file is available.
-
-        :param filename: The file name to use when sending the file to
-                         the client.
-        """
-        loc = self.obj.bucket.location.uri
-        bucket = self.aws_connection.get_bucket(loc)
-        element = Key(bucket)
-        element.key = self.obj.file.uri
-        return redirect(element.generate_url(expires_in=300))
-
-    def save(self, incoming_stream, size=None, chunk_size=None):
-
-        # Retrieve the bucket name
-        loc = self.obj.bucket.location.uri
-        bucket = self.aws_connection.get_bucket(loc)
-        element = Key(bucket)
-        element.key = "".join((incoming_stream.name, self.make_path()))
-
-        chunk_size = chunk_size or 1024 * 64
-        m = hashlib.md5()
-        bytes_written = 0
-
-        while 1:
-            chunk = incoming_stream.read(chunk_size)
-            if not chunk:
-                break
-            element.set_contents_from_string(chunk, replace=False)
-            bytes_written += len(chunk)
-            m.update(chunk)
-        element.close()
-        s3_uri = "".join(("s3:", bucket.name, ":", element.key))
-        self.obj.file.uri = s3_uri
-        return s3_uri, bytes_written, m.hexdigest()
-
-    def make_path_vid(self, vid):
-        """Make path to file in a given storage location."""
-        return join(
-            str(self.obj.bucket_id),
-            str(vid),
-        )
